'done test and frontend implimentation of multidot numpy'

+++ ivy_tests/test_ivy/test_frontends/test_numpy/test_linalg/test_matrix_and_vector_products.py
@@ -366,7 +366,7 @@ def test_numpy_multi_dot(
 @handle_frontend_test(
     fn_tree="numpy.outer",
     dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
+        available_dtypes=helpers.get_dtypes("float"),
         min_value=-10,
         max_value=10,
         num_arrays=2,
